refactor(stella-plugin): split manifest.rs's tests into their own file
manifest.rs sat at 1493 of its 1500-line ceiling — seven lines, and the
next person to touch it would have paid for a seam inside a change about
something else. #5208 already paid it once: a fourteen-line rule pushed
the file to 1506 and reddened CI, and the fix was chosen under a ceiling
rather than on design grounds.

**The tests moved, not the validation rules.** #5228 proposes splitting
`PluginManifest::validate` by grant type, and lists as its first
constraint that behaviour must not change — several tests match on
*which* error a manifest breaking two rules reports first. Relocating
rules across module boundaries is exactly what puts that ordering at
risk, for about 200 lines of relief. Moving the test module cannot
change behaviour at all and yields 676.

So the validate seam is still worth cutting on its own merits, and it can
now be cut deliberately rather than under a ceiling — which is the state
#5228 wanted. It stays open for that.

manifest.rs 1493 -> 817, with 683 lines of headroom. No test was renamed,
reordered or dropped: 30 before, 30 after, and `cargo test -p
stella-plugin` passes every suite. No baseline entry: the file is not
grandfathered and must not become so.

Refs #5228
